requirements.txt
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
click==8.2.0
cssselect2==0.8.0
Flask==3.1.0
fonttools==4.58.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
pip==24.2
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
gunicorn==22.0.0
